# openapi-spec.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /articles:
    post:
      operationId: ArticlesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleEntity'
      tags: &ref_0
        - articles
    get:
      operationId: ArticlesController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleEntity'
      tags: *ref_0
  /articles/drafts:
    get:
      operationId: ArticlesController_findDrafts
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleEntity'
      tags: *ref_0
  /articles/{id}:
    get:
      operationId: ArticlesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleEntity'
      tags: *ref_0
    patch:
      operationId: ArticlesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleEntity'
      tags: *ref_0
    delete:
      operationId: ArticlesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleEntity'
      tags: *ref_0
  /users:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
      tags: &ref_1
        - users
    get:
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserEntity'
      tags: *ref_1
      security:
        - bearer: []
  /users/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
      tags: *ref_1
      security:
        - bearer: []
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
      tags: *ref_1
      security:
        - bearer: []
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
      tags: *ref_1
      security:
        - bearer: []
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEntity'
      tags:
        - auth
info:
  title: Median
  description: The Median API description
  version: '0.1'
  contact: {}
tags:
  - name: example
    description: ''
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateArticleDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        body:
          type: string
        published:
          type: boolean
          default: false
      required:
        - title
        - body
    UserEntity:
      type: object
      properties:
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        name:
          type: string
        email:
          type: string
      required:
        - id
        - createdAt
        - updatedAt
        - name
        - email
    ArticleEntity:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
          nullable: true
        body:
          type: string
        published:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        authorId:
          type: number
          nullable: true
        author:
          $ref: '#/components/schemas/UserEntity'
      required:
        - id
        - title
        - body
        - published
        - createdAt
        - updatedAt
    UpdateArticleDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        body:
          type: string
        published:
          type: boolean
          default: false
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthEntity:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken